+++ CHANGELOG.md
@@ -7,9 +7,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ❤️ **Donate:** Enjoying MagicMirror²? [Please consider a donation!](https://magicmirror.builders/donate) With your help we can continue to improve the MagicMirror² core.
 
-## [2.7.x] - 2019-04-13
+## [2.8.0] - Unreleased
 
-default module calendar - added config option nextDaysRelative
+*This release is scheduled to be released on 2019-04-01.*
+
+### Added
+- Option to show event location in calendar
+
+- Finnish translation for "Feels" and "Weeks"
+
+- Russian translation for “Feels”
+
+- default module calendar - added config option nextDaysRelative
+
+### Updated
+
+### Fixed
+- Handle SIGTERM messages
+- Fixes sliceMultiDayEvents so it respects maximumNumberOfDays
 
 ## [2.7.1] - 2019-04-02